ejemplo de clase con autenticación

+++ src/main/java/net/ausiasmarch/foxforumserver/service/ReplyService.java
@@ -7,21 +7,30 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import jakarta.servlet.http.HttpServletRequest;
 import net.ausiasmarch.foxforumserver.entity.ReplyEntity;
+import net.ausiasmarch.foxforumserver.entity.UserEntity;
 import net.ausiasmarch.foxforumserver.exception.ResourceNotFoundException;
 import net.ausiasmarch.foxforumserver.helper.DataGenerationHelper;
 import net.ausiasmarch.foxforumserver.repository.ReplyRepository;
+import net.ausiasmarch.foxforumserver.repository.UserRepository;
 
 @Service
 public class ReplyService {
     @Autowired
     ReplyRepository oReplyRepository;
 
+    @Autowired
+    UserRepository oUserRepository;
+
+    @Autowired
+    HttpServletRequest oRequest;
+
     public ReplyEntity get(Long id) {
         return oReplyRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Reply not found"));
     }
 
-    public Long create(ReplyEntity oReplyEntity) {
+    public Long create(ReplyEntity oReplyEntity) {    
         oReplyEntity.setId(null);
         return oReplyRepository.save(oReplyEntity).getId();
     }
@@ -31,8 +40,25 @@ public class ReplyService {
     }
 
     public Long delete(Long id) {
-        oReplyRepository.deleteById(id);
-        return id;
+
+        // ejemplo de clase:
+        // si el usuario es user: si el reply es suyo entonces puede borrarlo
+        // si el usuario es admin: puede borrarlo
+
+        ReplyEntity oReplyEntity = oReplyRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Reply not found"));
+        String username = oRequest.getAttribute("username").toString();
+        UserEntity oUserEntity = oUserRepository.findByUsername(username).orElseThrow(() -> new ResourceNotFoundException("User not found"));
+        if (oUserEntity.getRole()) {
+            oReplyRepository.deleteById(id);
+            return id;
+        } else {
+            if (oReplyEntity.getUser().getId() == oUserEntity.getId()) {
+                oReplyRepository.deleteById(id);
+                return id;
+            } else {
+                throw new ResourceNotFoundException("Unauthorized");
+            }
+        }
     }
 
     public Page<ReplyEntity> getPage(Pageable oPageable, Long userId, Long threadId) {
